+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Leaflet Changelog
  * Added **WMS support** (L.TileLayer.WMS), currently EPSG:3857 only.
  * `L.Circle` is now zoom-dependent (with radius in meters); circle of a permanent size is now called `L.CircleMarker`.
  * Disabled zoom animation on Android by default because it's buggy on some devices (will be enabled back when it's stable enough). [#32](https://github.com/CloudMade/Leaflet/issues/32)
+ * Added Map `mouseover`, `mousemove` and `mouseout` events.
  * Added `setLatLng` method to `L.Marker`.
  * Added `maxZoom` argument to `map.locateAndSetView` method.
  * Improved geolocation error handling: better error messages, explicit timeout, set world view on locateAndSetView failure. [#61](https://github.com/CloudMade/Leaflet/issues/61)